% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{add_zt}
\alias{add_zt}
\title{Add Zeitgeber Time (ZT)}
\usage{
add_zt(df, datetime_col, lights_on_hour)
}
\arguments{
\item{df}{A \code{data.frame}}

\item{datetime_col}{The name of the column containing the datetime values to be shifted}

\item{lights_on_hour}{A numeric value between 1 and 23 representing the hour at which lights are turned on,
and which should become the new midnight (00:00:00)}
}
\value{
A dataframe with the new \code{zt} column
}
\description{
This function adds a new column \code{zt} to the input \code{data.frame}, which contains
the datetime values in \code{datetime_col} shifted so that \code{lights_on_hour} becomes the new midnight (00:00:00).
}
\examples{
df <- data.frame(
  datetime_col = seq(from = as.POSIXct("2023-07-20 00:00:00", tz = "UTC"),
                     to = as.POSIXct("2023-07-21 00:00:00", tz = "UTC"),
                     by = "hour"),
  value = runif(25)
)

lights_on_hour = 6
df <- add_zt(df, datetime_col, lights_on_hour)

}
\seealso{
\code{\link[=set_alignment]{set_alignment()}}, \code{\link[=bin_pellets_lightcycle]{bin_pellets_lightcycle()}}
}
