% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_sd.R
\name{backup_sd}
\alias{backup_sd}
\title{Backup SD card}
\usage{
backup_sd(sd_card = NULL, backup_dest = NULL)
}
\description{
Backup SD card
}
\details{
This function finds FED3 files on a selected folder (SD card) and stores them on a user-defined folder.
}
\seealso{
\code{fed3::find_files()}, \code{fed3::copy_files()}

This needs interactive session for user to choose folders
We assume interactive session likely thorugh Rstudio
The core of this function is a moving files function
It uses the pattern "FED"
It uses the date as MMDDYY to create folders on the backup location
}
