% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pellets.R
\name{bin_pellets_lightcycle}
\alias{bin_pellets_lightcycle}
\title{Bin Pellet Events According to Light Cycle}
\usage{
bin_pellets_lightcycle(
  data,
  time_col,
  lights_on_hour = 7,
  lights_off_hour = 19
)
}
\arguments{
\item{data}{A data frame that contains a datetime column and only pellet events.}

\item{time_col}{The datetime column in your data frame. You can use a bare column name.}

\item{lights_on_hour}{The hour (0-23) when the light period starts. Default is 7.}

\item{lights_off_hour}{The hour (0-23) when the light period ends. Default is 19.}
}
\value{
A \code{data.frame} grouped by the original grouping variables, date and light cycle period,
with an additional column \code{pellets} indicating the number of events in each period.
}
\description{
This function assigns each event to a light or dark period, then counts the number of events in each period.
It requires a data frame that contains a datetime column and only pellet events.

Because dates and light cycles are not aligned (dark phase often spans more than one date), it is recommended that users call \code{add_zt()} first and bin by \code{zt} instead of \code{datetime}.
}
\examples{

\dontrun{
# will have light cycle split by date
read_fed(path) \%>\% recalculate_pellets() \%>\% add_zt(datetime) \%>\% bin_pellets_lightcycle(datetime)
read_fed(path) \%>\% recalculate_pellets() \%>\% add_zt(datetime) \%>\% bin_pellets_lightcycle(datetime)
# zt date will only contain full light/dark periods
read_fed(path) \%>\% recalculate_pellets() \%>\% add_zt(datetime) \%>\% bin_pellets_lightcycle(zt)
}

}
\seealso{
\code{\link[=bin_pellets]{bin_pellets()}}, \code{\link[=add_zt]{add_zt()}}, \code{\link[=filter_pellets]{filter_pellets()}}, \code{\link[=recalculate_pellets]{recalculate_pellets()}}
}
