% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_files.R
\name{copy_files}
\alias{copy_files}
\title{Copy Files}
\usage{
copy_files(file_list = NULL, from_dir = NULL, dest_dir = NULL)
}
\description{
This function copies files from a path to a new folder at destiny. This function will also delete empty files, but it will not delete very small files.
}
\seealso{
`[find_files()]`
}
