% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fed.R
\name{deduplicate_datetime}
\alias{deduplicate_datetime}
\title{Check Duplicated}
\usage{
deduplicate_datetime(
  data,
  method = "offset",
  offset = "0.1 sec",
  reset_counts = FALSE,
  reset_columns = c("Pellet_Count", "Left_Poke_Count", "Right_Poke_Count")
)
}
\arguments{
\item{data}{FED data frame as read by \code{read_fed()}}

\item{method}{The method to deduplicate the identical timestamps (default is 'offset'). \code{method} must be one of 'keep_first', 'keep_last', 'remove', or 'offset', or 'interpolate'.}

\item{offset}{The offset to be added to duplicate datetimes (e.g., '0.1 sec').}

\item{reset_counts}{whether to reset the \code{reset_columns} or not (default = FALSE).}

\item{reset_columns}{The columns to be reset if \code{reset_counts = TRUE}.}
}
\description{
The precision of the FED clock is in seconds.
It might happen that two events have the same exact timestamp.
This helper function is used to check for duplicated entries and deduplicate them using methods.
This function is internally called by \code{read_fed()} and the user is advised to first try several examples before modifying \code{deduplicate_method} in \code{read_fed()}
}
\examples{
# data contains datetimes that will fail to parse and are duplicated
fed3::duplicate_test_data
fed3:::deduplicate_datetime(duplicate_test_data, method = 'keep_first')
fed3:::deduplicate_datetime(duplicate_test_data, method = 'offset', offset = "1 sec")
}
