% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fed3-package.R
\docType{package}
\name{fed3-package}
\alias{fed3}
\alias{fed3-package}
\title{fed3: An R package to read and analyze data from FED3 devices}
\description{
An R package to read and analyze data from FED3 devices
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matiasandina/fed3}
  \item Report bugs at \url{https://github.com/matiasandina/fed3/issues}
}

}
\author{
\strong{Maintainer}: Matias Andina \email{matiasandina@gmail.com} (0000-0002-1996-2539)

}
\keyword{internal}
