% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fed_col_types.R
\name{fed_col_types}
\alias{fed_col_types}
\title{Fed Column Types}
\usage{
fed_col_types(lib_version = NULL)
}
\arguments{
\item{version}{character version of FED_library used to collect data}
}
\description{
This function provides the column classes that will be used to read data from FEDs. Because the different library version might have different column types, this function returns the types according to the \code{lib_version}.
}
