% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_files.R
\name{find_files}
\alias{find_files}
\title{Find Files}
\usage{
find_files(target_dates = NULL, directory = NULL)
}
\arguments{
\item{target_dates}{The dates you want to retrieve from \code{directory}}

\item{directory}{The directory path where FED3 data is located (default = NULL)}
}
\description{
This function finds all FED files using target dates
}
\seealso{
\code{fed3::copy_files()}
}
