% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_files.R
\name{is.Date}
\alias{is.Date}
\title{Nice function to check if things are dates
We don't want to coerce it, it can lead to a lot of errors}
\usage{
is.Date(x)
}
\description{
Nice function to check if things are dates
We don't want to coerce it, it can lead to a lot of errors
}
