% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_files.R
\name{is_date}
\alias{is_date}
\title{Is Date}
\usage{
is_date(x)
}
\value{
\code{boolean}
}
\description{
Helper function to check if things are of class \code{Date} or \code{POSIXt}.
Equivalent to \code{function(x) {lubridate::is.Date(x) | lubridate::is.POSIXt(x)}}
}
