% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lights.R
\name{make_lights}
\alias{make_lights}
\title{Make Lights}
\usage{
make_lights(params, df)
}
\arguments{
\item{params}{A `list` that contains a `lights` `vector` with the ON-OFF light cycle}

\item{df}{A `data.frame` containing a `datetime` column that will provide the ranges of the experimental data}
}
\description{
This function will make light shading pattern for ggplot plots
}
