% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_files.R
\name{move_and_clean_files}
\alias{move_and_clean_files}
\title{Move and Clean FED3 Files}
\usage{
move_and_clean_files(files, inputdir, remove_small_files = TRUE)
}
\arguments{
\item{files}{Vector of FED3 file paths to process.}

\item{inputdir}{Base directory from which files are processed.}

\item{remove_small_files}{Logical indicating whether to remove small files.}
}
\value{
A list of processed file paths.
}
\description{
Removes files smaller than 300 bytes and moves remaining files to a specified subdirectory ('FED_data/FED_data').
}
