% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fed.R
\name{read_fed}
\alias{read_fed}
\title{"Read Fed Files"}
\usage{
read_fed(filename, lib_version = NULL, deduplicate_method = "offset")
}
\arguments{
\item{filename}{file path of csv file produced by FED3}
}
\description{
"Read Fed Files"
}
\details{
This function will read the raw data with column specifications (see \code{fed_col_types()}). It will append the \code{datetime} by formatting the FED's clock datetime as \verb{yyyy-mm-dd HH:MM:SS}, it will also extract \code{year}, \code{month}, and \code{day} from that date. It will parse the session from the \code{filename} and append it to the resulting \code{data.frame}.
}
