% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pellets.R
\name{recalculate_pellets}
\alias{recalculate_pellets}
\title{This function recalculates pellets if given a FED_data `data.frame` that contains an identifier column `ID`
The basic thought behind it is that one animal can receive more than one FED device
This might happen due to the experiment design or because a FED needed to be replaced during the experiment}
\usage{
recalculate_pellets(df)
}
\description{
This function recalculates pellets if given a FED_data `data.frame` that contains an identifier column `ID`
The basic thought behind it is that one animal can receive more than one FED device
This might happen due to the experiment design or because a FED needed to be replaced during the experiment
}
