% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lights.R
\docType{data}
\name{StatLights}
\alias{StatLights}
\title{StatLights}
\format{
An object of class \code{StatLights} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatLights
}
\arguments{
\item{lights_on_hour}{An integer representing the hour of the day (0-23) when the lights turn on.}

\item{lights_off_hour}{An integer representing the hour of the day (0-23) when the lights turn off.}

\item{fill}{A string representing the color of the shading.}

\item{alpha}{A numeric representing the transparency of the shading, between 0 (transparent) and 1 (opaque).}
}
\value{
A ggplot2 layer to be added to a ggplot object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
A custom ggplot2 extension to shade areas in a plot according to provided light hours.
It shades from lights off to lights on. Usage through \code{\link[=geom_lights]{geom_lights()}} or \code{\link[=stat_ligths]{stat_ligths()}}
}
\examples{
\dontrun{
ggplot(df, aes(x = Pi_Time, y = Pellet_Count)) +
  stat_lights(fill = "grey80", alpha = 0.5, lights_on_hour = 7, lights_off_hour = 19) +
  geom_line()

}
}
\keyword{datasets}
