% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_files.R
\name{tidy_fed3_files}
\alias{tidy_fed3_files}
\title{Tidy FED3 Files}
\usage{
tidy_fed3_files(inputdir = NULL, remove_small_files = TRUE)
}
\arguments{
\item{inputdir}{Optional directory to search for FED3 files.
If not provided, a directory selection dialog will open.}

\item{remove_small_files}{Logical indicating whether files smaller than 300 bytes should be removed.}
}
\description{
Handles directory selection, file finding, and file management operations for FED3 files.
}
\examples{
\dontrun{
tidy_fed3_files()
}
}
