% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pi_time.R
\name{use_pi_time}
\alias{use_pi_time}
\title{Adjust datetime in FED data using Pi_Time from FEDWatcher}
\usage{
use_pi_time(fed_df, fw_df)
}
\arguments{
\item{fed_df}{A data frame, the result of fed3::read_fed(...).}

\item{fw_df}{A data frame, the result of fed3::read_fed(..., lib_version = "fw").}
}
\value{
A data frame identical to `fed_df` but with `Pi_Time` adjusted based
on the linear model trained on `fw_df`.
}
\description{
This function adjusts the time in `fed_df` dataset
using a linear model trained on data coming from FEDWatcher (where `Pi_Time` is not NA).
You can use this function even if `fed_df` and `fw_df` do not fully overlap
in time.
}
\examples{
\dontrun{
fed_df <- fed3::read_fed(...)
fw_df <- fed3::read_fed(..., lib_version = "fw")
fed_df_adjusted <- use_pi_time(fed_df, fw_df)
}
}
