% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fed_col_types.R
\name{fed_col_types}
\alias{fed_col_types}
\title{"Fed Column Types"}
\usage{
fed_col_types(lib_version = NULL)
}
\arguments{
\item{version}{character version of FED_library used to collect data}
}
\description{
"Fed Column Types"
}
