% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lights.R
\name{make_lights}
\alias{make_lights}
\title{Make Lights}
\usage{
make_lights(params, df)
}
\arguments{
\item{params}{A \code{list} that contains a \code{lights} \code{vector} with the ON-OFF light cycle}

\item{df}{A \code{data.frame} containing a \code{datetime} column that will provide the ranges of the experimental data}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function will make light shading pattern for ggplot plots
}
